web/requirements.txt
# web/requirements.txt
flask
grpcio
grpcio-tools